server/pyproject.toml
[tool.poetry]
name = "cms-cloud-manager-frontend"
version = "0.1.0"
description = ""
authors = ["CMS Cloud Manager"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
flask-cors = "^5.0.1"
bcrypt = "^4.3.0"
flask-jwt-extended = "^4.7.1"
cryptography = "^44.0.2"
dotenv = "^0.9.9"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
